+++ Makefile.am
@@ -5,13 +5,18 @@
 
 ACLOCAL_AMFLAGS = -I build-aux/m4
 SUBDIRS = src
-if ENABLE_MAN
-SUBDIRS += doc/man
-endif
+#if ENABLE_MAN
+#SUBDIRS += doc/man
+#endif
 .PHONY: deploy FORCE
 
 GZIP_ENV="-9n"
 export PYTHONPATH
+
+#if BUILD_BITCOIN_LIBS
+#pkgconfigdir = $(libdir)/pkgconfig
+#pkgconfig_DATA = libbitcoinconsensus.pc
+#endif
 
 BITCOIND_BIN=$(top_builddir)/src/$(BITCOIN_DAEMON_NAME)$(EXEEXT)
 BITCOIN_QT_BIN=$(top_builddir)/src/qt/$(BITCOIN_GUI_NAME)$(EXEEXT)
