+++ src/compat/glibc_sanity.cpp
@@ -3,7 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #if defined(HAVE_CONFIG_H)
-#include "config/bitcoin-config.h"
+#include "config/yacoin-config.h"
 #endif
 
 #include <cstddef>
