+++ src/leveldb/Makefile
@@ -16,7 +16,7 @@ OPT ?= -O2 -DNDEBUG
 
 # detect what platform we're building on
 $(shell CC="$(CC)" CXX="$(CXX)" TARGET_OS="$(TARGET_OS)" \
-    sh ./build_detect_platform build_config.mk ./)
+    ./build_detect_platform build_config.mk ./)
 # this file is generated by the previous line to set build flags and sources
 include build_config.mk
